delete unused code. refactor code

+++ client/src/pages/home/components/product-card/ProductCard.jsx
@@ -1,28 +1,35 @@
-import { motion } from 'framer-motion';
 import { useEffect, useState } from 'react';
-import { FiShoppingBag, FiSearch, FiHeart } from 'react-icons/fi';
-
-import { productCardAnimation } from '../../helpers/motions-utils';
-import { ButtonAnimated } from '../';
-import { Wrapper, Image, ButtonsWrapper } from './ProductCard.styled';
 import { useDispatch } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { FiShoppingBag, FiSearch, FiHeart } from 'react-icons/fi';
+import { motion } from 'framer-motion';
+
 import {
 	addToCart,
 	addToFavourites,
 	removeFromCart,
 	removeFromFavourites,
-} from '../../redux/actions';
-import { routeToProductPage } from '../../router/routes';
-import { useNavigate } from 'react-router-dom';
+} from '../../../../redux/actions';
+
+import { LOGIN_PAGE_ROUTE, routeToProductPage } from '../../../../router/routes';
+import { productCardAnimation } from '../../../../helpers/motions-utils';
+import { ButtonAnimated } from '../../../../shared';
+import { ProductCardSkeleton } from './ProductCardSkeleton';
+
+import { Wrapper, Image, ButtonsWrapper } from './ProductCard.styled';
 
 export const ProductCard = (props) => {
-	const { product, image, index, title, checkIsInCart, checkIsFavourite } = props;
+	const { product, image, index, isAuth, checkIsInCart, checkIsFavourite } = props;
 	const [isInCart, setIsInCart] = useState(false);
 	const [isFavourite, setIsFavourite] = useState(false);
+	const [isImageLoaded, setIsImageLoaded] = useState(false);
 	const dispatch = useDispatch();
 	const navigate = useNavigate();
 
 	const handleToggleToCart = () => {
+		// FIXME: change to info popup with link - you need to login before
+		if (!isAuth) return navigate(LOGIN_PAGE_ROUTE);
+
 		if (isInCart) {
 			dispatch(removeFromCart(product));
 		} else {
@@ -36,6 +43,9 @@ export const ProductCard = (props) => {
 	};
 
 	const handleToggleToFavourite = () => {
+		// FIXME: change to info popup with link - you need to login before
+		if (!isAuth) return navigate(LOGIN_PAGE_ROUTE);
+
 		if (isFavourite) {
 			dispatch(removeFromFavourites(product));
 		} else {
@@ -50,11 +60,18 @@ export const ProductCard = (props) => {
 
 		setIsInCart(isAlreadyInCart);
 		setIsFavourite(isAlreadyFavourite);
+		// eslint-disable-next-line
 	}, []);
 
 	return (
 		<Wrapper as={motion.div} custom={index} {...productCardAnimation}>
-			<Image src={image} alt={title} />
+			<Image
+				style={isImageLoaded ? {} : { display: 'none' }}
+				src={image}
+				alt={product.title}
+				onLoad={() => setIsImageLoaded(true)}
+			/>
+			<ProductCardSkeleton style={isImageLoaded ? { display: 'none' } : {}} />
 			<ButtonsWrapper>
 				<ButtonAnimated onClick={handleToggleToFavourite} isActive={isFavourite}>
 					<FiHeart />
